requirements.txt
# requirements.txt

# Core CIE & ADK
google-adk==0.5.0
google-cloud-firestore
google-cloud-storage
google-cloud-translate
google-cloud-aiplatform
google-cloud-tasks
python-dotenv
requests
beautifulsoup4
litellm
deprecated

# --- Web Server Stack (PINNED FOR STABILITY) ---
Flask[async]==2.3.3
Werkzeug==2.3.8
gunicorn==21.2.0
uvicorn